Refactor/week3/point

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Request.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 public class PointV1Request {
     public record PointChargeRequest(
             @NotNull
-            int amount
+            Long amount
     ) {
 
         public PointCommand.Charge toCommand(Long userId) {
